Backoff and retry subscription failures
AWS IoT limits the rate that subscriptions can be made so if something
causes a lot of devices to reconnect at the same time, the subscription
step is what fails first.

Prior to this change, the subscription failure would stop the Connection
GenServer. It would immediately respawn and try again. This ended up
increasing the chance that a subscription would fail.

The backoff used here is the same as the reconnect backoff for
simplicity.

As another justification for retrying subscriptions when they fail vs.
closing the connection and trying again, it seems unlikely that
reconnecting would do any better in fixing the problem. Therefore, do
the least resource/bandwidth intensive thing and retry subscribing. If
the socket gets disconnected, it will break this retry loop.