Cap the max reconnect rate by delaying before the first connect
If the Connection GenServer crashes or stops due to an error it gets
restarted immediately. This is problematic when remaking the connection
immediately doesn't allow the issue to resolve itself. A trivial way of
relieving the issue is to add a short manditory reconnect delay. Since
the delay is mandatory, it can't be long. Given typical connect times,
the default of 1 second probably won't be noticed. 50% jitter is added.

The delay can be changed using the `:first_connect_delay` option.